I chose to heavily refactor this.  The original had a hardcoded decision tree based on potentially flexible "types".  This is a smell as any charge to these plans would require a re-deployment, and they are likely to change or require additional plans.
I've implemented a flexible price plan that has the concept of
break points on both number of phones to price and number of phones
to discount.  The old hardcoded plans are mapped to this using
the appropriate break points and the initializer extend to also
accept the break points as arguments.

In the future this would allow the (obvious) extension to allow
price plans to be described in the database and the breakpoints
simply scoped in.

There was also certain - potentially edge - cases that werent explicitly
handled and, although the default action mave have been correct, the
outcome was ambiguous (for example an individual plan with more than one
phone, or a family plan with only 1 phone)

I haven't changed the tests but they should really have explicit
determination of what these edge cases should do also.

As far as the original code is concerned, I would have short cut the
cost if/else ladder with explicit returns of the value - the method is
too long to be obvious that the value is falling through.
The formula for each type could've been pulled in to separate methods
to localize each type/calculation but I would've refactored this
completely before needing to do this.